Refactor code and reduce duplication
Puts all wrapper logic in one place and dedups the methods that had been
copied. Also removes wrapper interfaces, since they can't participate in
deserialization.
