+++ src/generator/Generator.cs
@@ -46,25 +46,6 @@ namespace Serde
     [Generator]
     public partial class SerdeImplRoslynGenerator : IIncrementalGenerator
     {
-        internal static string? SerdeBuiltInName(SpecialType specialType) => specialType switch
-        {
-            SpecialType.System_String => "String",
-            SpecialType.System_Boolean => "Bool",
-            SpecialType.System_Char => "Char",
-            SpecialType.System_Byte => "Byte",
-            SpecialType.System_UInt16 => "U16",
-            SpecialType.System_UInt32 => "U32",
-            SpecialType.System_UInt64 => "U64",
-            SpecialType.System_SByte => "SByte",
-            SpecialType.System_Int16 => "I16",
-            SpecialType.System_Int32 => "I32",
-            SpecialType.System_Int64 => "I64",
-            SpecialType.System_Single => "Float",
-            SpecialType.System_Double => "Double",
-            SpecialType.System_Decimal => "Decimal",
-            _ => null
-        };
-
         /// <inheritdoc />
         public void Initialize(IncrementalGeneratorInitializationContext context)
         {
@@ -129,7 +110,7 @@ namespace Serde
 
                     if (typeDecl.IsKind(SyntaxKind.EnumDeclaration))
                     {
-                        GenerateEnumWrapper(
+                        Wrappers.GenerateEnumWrapper(
                             typeDecl,
                             attrCtx.SemanticModel,
                             generationContext);
