+++ src/generator/HelpersAndUtilities/SymbolUtilities.cs
