+++ src/generator/SerdeTypeInfoGenerator.cs
