+++ src/generator/WellKnownTypes.cs
@@ -15,8 +15,6 @@ namespace Serde
         ImmutableArray_1,
         List_1,
         Dictionary_2,
-        IDictionary_2,
-        IReadOnlyDictionary_2,
     }
 
     [Closed]
@@ -42,25 +40,6 @@ namespace Serde
                 }
             }
             return null;
-        }
-
-        public static ImmutableArray<INamedTypeSymbol> GetAvailableInterfacesInOrder(GeneratorExecutionContext context)
-        {
-            var builder = ImmutableArray.CreateBuilder<INamedTypeSymbol>();
-            // Order matters here -- checking for interface implementation happens in order and
-            // bails out early on the first success
-            var candidates = new[] {
-                WellKnownType.IDictionary_2,
-                WellKnownType.IReadOnlyDictionary_2
-            };
-            foreach (var candidate in candidates)
-            {
-                if (context.Compilation.GetTypeByMetadataName(candidate.GetFQN()) is {} type)
-                {
-                    builder.Add(type);
-                }
-            }
-            return builder.ToImmutable();
         }
 
         internal static string GetName(this WellKnownAttribute wk) => wk switch
@@ -89,8 +68,6 @@ namespace Serde
             "ImmutableArray`1" => WellKnownType.ImmutableArray_1,
             "List`1" => WellKnownType.List_1,
             "Dictionary`2" => WellKnownType.Dictionary_2,
-            "IDictionary`2" => WellKnownType.IDictionary_2,
-            "IReadOnlyDictionary`2" => WellKnownType.IReadOnlyDictionary_2,
              _ => null
         };
 
@@ -99,8 +76,6 @@ namespace Serde
             WellKnownType.ImmutableArray_1 => "System.Collections.Immutable.ImmutableArray`1",
             WellKnownType.List_1 => "System.Collections.Generic.List`1",
             WellKnownType.Dictionary_2 => "System.Collections.Generic.Dictionary`2",
-            WellKnownType.IDictionary_2 => "System.Collections.Generic.IDictionary`2",
-            WellKnownType.IReadOnlyDictionary_2 => "System.Collections.Generic.IReadOnlyDictionary`2",
             _ => throw ExceptionUtilities.Unreachable
         };
 
@@ -111,8 +86,6 @@ namespace Serde
                 WellKnownType.ImmutableArray_1 => "ImmutableArrayWrap",
                 WellKnownType.List_1 => "ListWrap",
                 WellKnownType.Dictionary_2 => "DictWrap",
-                WellKnownType.IDictionary_2 => "IDictWrap",
-                WellKnownType.IReadOnlyDictionary_2 => "IRODictWrap",
                 _ => throw ExceptionUtilities.Unreachable
             };
             return baseName + "." + usage.GetImplName();
