+++ src/serde/Wrappers.Dictionary.cs
@@ -1,7 +1,4 @@
 using System.Collections.Generic;
-using System.ComponentModel;
-using System.ComponentModel.DataAnnotations;
-using System.Net.Http.Headers;
 
 namespace Serde
 {
@@ -63,53 +60,6 @@ namespace Serde
                     throw new InvalidDeserializeValueException($"Expected {size} items, found {dict.Count}");
                 }
                 return dict;
-            }
-        }
-    }
-
-    public static class IDictWrap
-    {
-        public readonly record struct SerializeImpl<TKey, TKeyWrap, TValue, TValueWrap> : ISerialize<IDictionary<TKey, TValue>>
-            where TKey : notnull
-            where TKeyWrap : struct, ISerialize<TKey>
-            where TValueWrap : struct, ISerialize<TValue>
-        {
-            public void Serialize(IDictionary<TKey, TValue> value, ISerializer serializer)
-            {
-                var typeInfo = DictSerdeTypeInfo<TKey, TValue>.TypeInfo;
-                var sd = serializer.SerializeCollection(typeInfo, value.Count);
-                foreach (var (k, v) in value)
-                {
-                    sd.SerializeElement(k, default(TKeyWrap));
-                    sd.SerializeElement(v, default(TValueWrap));
-                }
-                sd.End(typeInfo);
-            }
-        }
-    }
-
-    public static class IRODictWrap
-    {
-        private static readonly TypeInfo s_typeInfo = TypeInfo.Create(
-            typeof(IReadOnlyDictionary<,>).Name,
-            TypeInfo.TypeKind.Dictionary,
-            []);
-
-        public readonly record struct SerializeImpl<TKey, TKeyWrap, TValue, TValueWrap> : ISerialize<IReadOnlyDictionary<TKey, TValue>>
-            where TKey : notnull
-            where TKeyWrap : struct, ISerialize<TKey>
-            where TValueWrap : struct, ISerialize<TValue>
-        {
-            public void Serialize(IReadOnlyDictionary<TKey, TValue> value, ISerializer serializer)
-            {
-                var typeInfo = s_typeInfo;
-                var sd = serializer.SerializeCollection(typeInfo, value.Count);
-                foreach (var (k, v) in value)
-                {
-                    sd.SerializeElement(k, default(TKeyWrap));
-                    sd.SerializeElement(v, default(TValueWrap));
-                }
-                sd.End(typeInfo);
             }
         }
     }
