+++ test/Serde.Generation.Test/SerializeTests.cs
@@ -292,90 +292,6 @@ partial class C2
         }
 
         [Fact]
-        public Task IDictionaryImplGenerate()
-        {
-            var src = @"
-using System;
-using System.Collections;
-using System.Collections.Generic;
-using Serde;
-
-record R(Dictionary<string, int> D) : IDictionary<string, int>
-{
-    public int this[string key] { get => ((IDictionary<string, int>)D)[key]; set => ((IDictionary<string, int>)D)[key] = value; }
-
-    public ICollection<string> Keys => ((IDictionary<string, int>)D).Keys;
-
-    public ICollection<int> Values => ((IDictionary<string, int>)D).Values;
-
-    public int Count => ((ICollection<KeyValuePair<string, int>>)D).Count;
-
-    public bool IsReadOnly => ((ICollection<KeyValuePair<string, int>>)D).IsReadOnly;
-
-    public void Add(string key, int value)
-    {
-        ((IDictionary<string, int>)D).Add(key, value);
-    }
-
-    public void Add(KeyValuePair<string, int> item)
-    {
-        ((ICollection<KeyValuePair<string, int>>)D).Add(item);
-    }
-
-    public void Clear()
-    {
-        ((ICollection<KeyValuePair<string, int>>)D).Clear();
-    }
-
-    public bool Contains(KeyValuePair<string, int> item)
-    {
-        return ((ICollection<KeyValuePair<string, int>>)D).Contains(item);
-    }
-
-    public bool ContainsKey(string key)
-    {
-        return ((IDictionary<string, int>)D).ContainsKey(key);
-    }
-
-    public void CopyTo(KeyValuePair<string, int>[] array, int arrayIndex)
-    {
-        ((ICollection<KeyValuePair<string, int>>)D).CopyTo(array, arrayIndex);
-    }
-
-    public IEnumerator<KeyValuePair<string, int>> GetEnumerator()
-    {
-        return ((IEnumerable<KeyValuePair<string, int>>)D).GetEnumerator();
-    }
-
-    public bool Remove(string key)
-    {
-        return ((IDictionary<string, int>)D).Remove(key);
-    }
-
-    public bool Remove(KeyValuePair<string, int> item)
-    {
-        return ((ICollection<KeyValuePair<string, int>>)D).Remove(item);
-    }
-
-    public bool TryGetValue(string key, out int value)
-    {
-        return ((IDictionary<string, int>)D).TryGetValue(key, out value);
-    }
-
-    IEnumerator IEnumerable.GetEnumerator()
-    {
-        return ((IEnumerable)D).GetEnumerator();
-    }
-}
-[GenerateSerialize]
-partial class C
-{
-    public R RDictionary;
-}";
-            return VerifySerialize(src);
-        }
-
-        [Fact]
         public Task ExplicitWrapper()
         {
             var src = @"
