+++ test/Serde.Generation.Test/test_output/DeserializeTests.EnumMember/C.IDeserialize.verified.cs
@@ -21,19 +21,19 @@ partial class C : Serde.IDeserialize<C>
             switch (_l_index_)
             {
                 case 0:
-                    _l_colorint = typeDeserialize.ReadValue<ColorInt, global::ColorIntWrap>(_l_index_);
+                    _l_colorint = typeDeserialize.ReadValue<ColorInt, ColorIntWrap>(_l_index_);
                     _r_assignedValid |= ((byte)1) << 0;
                     break;
                 case 1:
-                    _l_colorbyte = typeDeserialize.ReadValue<ColorByte, global::ColorByteWrap>(_l_index_);
+                    _l_colorbyte = typeDeserialize.ReadValue<ColorByte, ColorByteWrap>(_l_index_);
                     _r_assignedValid |= ((byte)1) << 1;
                     break;
                 case 2:
-                    _l_colorlong = typeDeserialize.ReadValue<ColorLong, global::ColorLongWrap>(_l_index_);
+                    _l_colorlong = typeDeserialize.ReadValue<ColorLong, ColorLongWrap>(_l_index_);
                     _r_assignedValid |= ((byte)1) << 2;
                     break;
                 case 3:
-                    _l_colorulong = typeDeserialize.ReadValue<ColorULong, global::ColorULongWrap>(_l_index_);
+                    _l_colorulong = typeDeserialize.ReadValue<ColorULong, ColorULongWrap>(_l_index_);
                     _r_assignedValid |= ((byte)1) << 3;
                     break;
                 case Serde.IDeserializeType.IndexNotFound:
