+++ test/Serde.Generation.Test/test_output/MemberFormatTests.EnumValues/S.IDeserialize.verified.cs
@@ -18,7 +18,7 @@ partial struct S : Serde.IDeserialize<S>
             switch (_l_index_)
             {
                 case 0:
-                    _l_e = typeDeserialize.ReadValue<ColorEnum, global::ColorEnumWrap>(_l_index_);
+                    _l_e = typeDeserialize.ReadValue<ColorEnum, ColorEnumWrap>(_l_index_);
                     _r_assignedValid |= ((byte)1) << 0;
                     break;
                 case Serde.IDeserializeType.IndexNotFound:
