+++ test/Serde.Generation.Test/test_output/MemberFormatTests.EnumValues/S.ISerialize.verified.cs
@@ -10,7 +10,7 @@ partial struct S : Serde.ISerialize<S>
     {
         var _l_typeInfo = SSerdeTypeInfo.TypeInfo;
         var type = serializer.SerializeType(_l_typeInfo);
-        type.SerializeField<ColorEnum, global::ColorEnumWrap>(_l_typeInfo, 0, value.E);
+        type.SerializeField<ColorEnum, ColorEnumWrap>(_l_typeInfo, 0, value.E);
         type.End();
     }
 }
