+++ test/Serde.Generation.Test/test_output/SerializeTests.NestedEnumWrapper/C.ISerialize.verified.cs
@@ -10,7 +10,7 @@ partial class C : Serde.ISerialize<C>
     {
         var _l_typeInfo = CSerdeTypeInfo.TypeInfo;
         var type = serializer.SerializeType(_l_typeInfo);
-        type.SerializeFieldIfNotNull<Rgb?, Serde.NullableWrap.SerializeImpl<Rgb, global::RgbWrap>>(_l_typeInfo, 0, value.ColorOpt);
+        type.SerializeFieldIfNotNull<Rgb?, Serde.NullableWrap.SerializeImpl<Rgb, RgbWrap>>(_l_typeInfo, 0, value.ColorOpt);
         type.End();
     }
 }
