--- test/Serde.Generation.Test/test_output/SerializeTests/IDictionaryImplGenerate#CSerdeTypeInfo.verified.cs
@@ -1,10 +1,0 @@
-﻿//HintName: CSerdeTypeInfo.cs
-internal static class CSerdeTypeInfo
-{
-    internal static readonly Serde.TypeInfo TypeInfo = Serde.TypeInfo.Create(
-        "C",
-        Serde.TypeInfo.TypeKind.CustomType,
-        new (string, System.Reflection.MemberInfo)[] {
-("rDictionary", typeof(C).GetField("RDictionary")!)
-    });
-}
